Finished fiwing bugs
