+++ client/src/components/MyProfile.js
@@ -4,6 +4,8 @@ import '../App.css';
 import Button from "@mui/material/Button";
 import EditDetailsModal from '../modals/EditDetailsModal';
 import EventItem from './EventItem';
+import { ToastContainer } from "react-toastify";
+
 
 function ProfileItem(props) {
   // These states store the original event data
@@ -376,6 +378,7 @@ function MyProfile(props) {
           }
         </div>
       </div>
+      <ToastContainer/>
     </div>
   )
 }
